Rename SSHPush task into SSHService
Executing action was moved to ActionExecutor tasks and this is
handling check-in and action chain resume for salt ssh clients now.
